Create Shadowing.java
Note that looper() declares an automatic variable x, which shadows the instance
variable x. So when x = m; is executed, it is the local variable x that is changed not the
instance field x. So getX() never returns 0. If you remove int x = 0; from looper(), it
will print 0 and end.